disable services for fhir versions below R4B
