+++ hapi-fhir-jpaserver-subscription/src/main/java/ca/uhn/fhir/jpa/subscription/match/config/SubscriptionProcessorConfig.java
@@ -19,10 +19,6 @@
  */
 package ca.uhn.fhir.jpa.subscription.match.config;
 
-import ca.uhn.fhir.context.FhirContext;
-import ca.uhn.fhir.interceptor.api.IInterceptorBroadcaster;
-import ca.uhn.fhir.jpa.api.dao.DaoRegistry;
-import ca.uhn.fhir.jpa.searchparam.matcher.SearchParamMatcher;
 import ca.uhn.fhir.jpa.subscription.channel.api.IChannelFactory;
 import ca.uhn.fhir.jpa.subscription.channel.subscription.SubscriptionChannelRegistry;
 import ca.uhn.fhir.jpa.subscription.channel.subscription.SubscriptionDeliveryChannelNamer;
@@ -37,16 +33,12 @@ import ca.uhn.fhir.jpa.subscription.match.matcher.matching.ISubscriptionMatcher;
 import ca.uhn.fhir.jpa.subscription.match.matcher.matching.InMemorySubscriptionMatcher;
 import ca.uhn.fhir.jpa.subscription.match.matcher.subscriber.MatchingQueueSubscriberLoader;
 import ca.uhn.fhir.jpa.subscription.match.matcher.subscriber.SubscriptionActivatingSubscriber;
-import ca.uhn.fhir.jpa.subscription.match.matcher.subscriber.SubscriptionMatchDeliverer;
 import ca.uhn.fhir.jpa.subscription.match.matcher.subscriber.SubscriptionMatchingSubscriber;
 import ca.uhn.fhir.jpa.subscription.match.matcher.subscriber.SubscriptionRegisteringSubscriber;
 import ca.uhn.fhir.jpa.subscription.match.registry.SubscriptionLoader;
 import ca.uhn.fhir.jpa.subscription.match.registry.SubscriptionRegistry;
 import ca.uhn.fhir.jpa.subscription.model.config.SubscriptionModelConfig;
-import ca.uhn.fhir.jpa.topic.SubscriptionTopicLoader;
-import ca.uhn.fhir.jpa.topic.SubscriptionTopicMatchingSubscriber;
-import ca.uhn.fhir.jpa.topic.SubscriptionTopicRegistry;
-import ca.uhn.fhir.jpa.topic.SubscriptionTopicSupport;
+import ca.uhn.fhir.jpa.topic.SubscriptionTopicConfig;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Import;
 import org.springframework.context.annotation.Primary;
@@ -56,22 +48,12 @@ import org.springframework.context.annotation.Scope;
  * This Spring config should be imported by a system that pulls messages off of the
  * matching queue for processing, and handles delivery
  */
-@Import(SubscriptionModelConfig.class)
+@Import({SubscriptionModelConfig.class, SubscriptionTopicConfig.class})
 public class SubscriptionProcessorConfig {
 
 	@Bean
 	public SubscriptionMatchingSubscriber subscriptionMatchingSubscriber() {
 		return new SubscriptionMatchingSubscriber();
-	}
-
-	@Bean
-	public SubscriptionMatchDeliverer subscriptionMatchDeliverer(FhirContext theFhirContext, IInterceptorBroadcaster theInterceptorBroadcaster, SubscriptionChannelRegistry theSubscriptionChannelRegistry) {
-		return new SubscriptionMatchDeliverer(theFhirContext, theInterceptorBroadcaster, theSubscriptionChannelRegistry);
-	}
-
-	@Bean
-	public SubscriptionTopicMatchingSubscriber subscriptionTopicMatchingSubscriber() {
-		return new SubscriptionTopicMatchingSubscriber();
 	}
 
 	@Bean
@@ -95,16 +77,6 @@ public class SubscriptionProcessorConfig {
 	}
 
 	@Bean
-	public SubscriptionTopicRegistry subscriptionTopicRegistry() {
-		return new SubscriptionTopicRegistry();
-	}
-
-	@Bean
-	public SubscriptionTopicSupport subscriptionTopicSupport(FhirContext theFhirContext, DaoRegistry theDaoRegistry, SearchParamMatcher theSearchParamMatcher) {
-		return new SubscriptionTopicSupport(theFhirContext, theDaoRegistry, theSearchParamMatcher);
-	}
-
-	@Bean
 	public SubscriptionDeliveryChannelNamer subscriptionDeliveryChannelNamer() {
 		return new SubscriptionDeliveryChannelNamer();
 	}
@@ -112,11 +84,6 @@ public class SubscriptionProcessorConfig {
 	@Bean
 	public SubscriptionLoader subscriptionLoader() {
 		return new SubscriptionLoader();
-	}
-
-	@Bean
-	public SubscriptionTopicLoader subscriptionTopicLoader() {
-		return new SubscriptionTopicLoader();
 	}
 
 	@Bean
